pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-video"
version = "0.1.0"
description = "MCP server for video intelligence - analyze any video with AI vision"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
keywords = ["mcp", "video", "ai", "llm", "claude", "vision", "youtube"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "mcp>=1.0.0",
    "yt-dlp>=2024.0.0",
    "opencv-python-headless>=4.8.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
whisper = ["openai-whisper>=20231117"]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
]

[project.urls]
Homepage = "https://github.com/ipythonist/mcp-video"
Repository = "https://github.com/ipythonist/mcp-video"
Issues = "https://github.com/ipythonist/mcp-video/issues"

[project.scripts]
mcp-video = "mcp_video.server:main"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_video"]
